chat_with_images_with_text/requirements.txt
langchain
langchain-community
langchain-ollama
streamlit
langchain-google-genai
langchain-openai
python-dotenv